requirements.txt
openstacksdk>=2.0
PyYAML>=6.0
python-openstackclient>=6.0
